modify : change multiply and 201802151 assignment code
